Don't hardcode the master branch when pushing to PUSH_REMOTE.
Instead, let git push whatever branches it is configured to push to that
remote.

This allows for branches other than master to be checked out and be pushed.

git has several configs for what branches to push, including
push.default, and remote.*.push.

This commit was sponsored by Trenton Cronholm on Patreon.