Support platforms without `hostname`
I just recently started using etckeeper on one machine (thanks, I like
it!), and then a few days ago set up a new Arch Linux machine and went
to run etckeeper on it, but it failed to init properly due to depending
on the `hostname` binary, which is no longer installed by default on
Arch Linux.

Concretely, /usr/bin/hostname comes from the inetutils package, which
used to be installed by default, being part of the base group, but is so
no longer, not being a dependency of the new base package. [1][2]

I figure that `cat /etc/hostname` is a decent alternative.

I license this patch under the terms of the GNU GPL, version 2 or later.

— Chris

[1]: https://www.archlinux.org/news/base-group-replaced-by-mandatory-base-package-manual-intervention-required/
[2]: https://www.reddit.com/r/archlinux/comments/de1er6/arch_linux_news_base_group_replaced_by_mandatory/f2qtzgj/