etck-commit: use rest of command line as message
Makes `commit` use the rest of the command line as message, i.e. not
just the first argument. This means that now all of the following work:

```
etckeeper commit -m'My message'
etckeeper commit -m 'My message'
etckeeper commit 'My message'
etckeeper commit My message
etckeeper commit -m My message
```

This was motivated by a lot of commits in my history that just had the
first word, because I got into the habit of not using quoting
characters, for whatever reason. Given that this is
backwards-compatible, there should be no reason not to include this.

Signed-off-by: martin f. krafft <madduck@madduck.net>