Improve sorting stability
In particular, make list-installed use a consistent sort order no matter
the locale. In post-install.d/50vcs-commit, an old and new output of that
are diffed, and if the locale varied between old and new being generated,
the diff could be longer than necessary, leading to more packages being
listed as changed than actually changed. Whether that could actually happen
in practice, I don't know.

In pre-commit.d/30store-metadata, there were two code blocks that both
tried to make the sort order stable. Combined them to make clearer what's
going on. And since LC_COLLATE beats LANG, only the former needs to be set.

The only other use of sort looks to be in post-install.d/50vcs-commit,
but there it is only being used in order to uniq, so stability does not
matter, so didn't change it there.

(It would be possible to set LC_COLLATE globally, but that risks somehow
affecting VCS internals or the editor run for etckeeper commit.)

This commit was sponsored by Ethan Aubin.